pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "binbook-poc"
version = "0.1.0"
description = "Python proof of concept for the BinBook raster-book format"
requires-python = ">=3.11"
dependencies = ["Pillow>=10"]

[dependency-groups]
dev = ["pytest>=8"]

[project.optional-dependencies]
dev = ["pytest>=8"]

[project.scripts]
binbook = "binbook.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
